Moved the path finding in the optimal orders class. changed class api
now the path is returned in a vector of tuples
also the matrix is of floats not doubles to improve performance
and cache utilization

+++ src/matrix.hpp
@@ -6,4 +6,4 @@
 
 #define NUMBER_CURRENCIES 33
 
-typedef double matrix[NUMBER_CURRENCIES][NUMBER_CURRENCIES];
+typedef float matrix[NUMBER_CURRENCIES][NUMBER_CURRENCIES];
